Division into smaller components
