+++ src/view/header/filter/filter-list.js
@@ -1,0 +1,7 @@
+export const createFilterListTemplate = () => {
+  return (
+    `<form class="trip-filters" action="#" method="get">
+      <button class="visually-hidden" type="submit">Accept filter</button>
+    </form>`
+  );
+};
