+++ src/view/header/filter/subcomponents/filter-item.js
@@ -1,0 +1,8 @@
+export const createFilterItemTemplate = (filterName) => {
+  return (
+    `<div class="trip-filters__filter">
+      <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
+      <label class="trip-filters__filter-label" for="filter-everything">${filterName}</label>
+    </div>`
+  );
+};
