+++ src/view/main/event_card/subcomponents/header/subcomponents/event-type-group.js
@@ -1,0 +1,7 @@
+export const createEventTypeGroupTemplate = (typeGroupName) => {
+  return (
+    `<fieldset class="event__type-group">
+      <legend class="visually-hidden">${typeGroupName}</legend>
+    </fieldset>`
+  );
+};
